Cleaning the code
