+++ src/CMakeLists.txt
@@ -1,19 +1,18 @@
 # source files of libibex-affine
-list (APPEND SRC ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine2_fAF2.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine2_fAF2.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine2.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine3_fAFFullI.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine3_fAFFullI.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineDomain.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVar.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVar.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineMatrix.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVector.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/function/ibex_AffineEval.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/function/ibex_AffineEval.h
-                ${CMAKE_CURRENT_SOURCE_DIR}/numeric/ibex_LinearizerAffine2.cpp
-                ${CMAKE_CURRENT_SOURCE_DIR}/numeric/ibex_LinearizerAffine2.h
+list (APPEND SRC 	${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine2_fAF2.cpp
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine2_fAF2.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine3_fAFFullI.cpp
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine3_fAFFullI.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineMain.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineDomain.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVar.cpp
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVar.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineMatrix.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_AffineVector.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/arithmetic/ibex_Affine.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/function/ibex_AffineEval.h
+	                ${CMAKE_CURRENT_SOURCE_DIR}/numeric/ibex_LinearizerAffine2.cpp
+	                ${CMAKE_CURRENT_SOURCE_DIR}/numeric/ibex_LinearizerAffine2.h
                 )
 
 # Create the target for libibex-affine
