--- src/arithmetic/ibex_Affine2.h
@@ -1,19 +1,0 @@
-/* ============================================================================
- * I B E X - Affine definition
- * ============================================================================
- * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
- *
- * Author(s)   : Jordan Ninin
- * Bug fixes   :
- * Created     : Nov 12, 2012
- * ---------------------------------------------------------------------------- */
-
-#ifndef IBEX_AFFINE2_H_
-#define IBEX_AFFINE2_H_
-
-// for the retro compatibility
-#include "ibex_Affine.h"
-
-#endif /* IBEX_Affine2_H_ */
-
