+++ src/arithmetic/ibex_Affine2_fAF2.cpp
@@ -3,10 +3,10 @@
  * ============================================================================
  * Copyright   : ENSTA Bretagne (FRANCE)
  * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
+ *               See the file LICENCE.
  *
  * Author(s)   : Jordan Ninin
- * Created     : Jul 16, 2013
+ * Created     : June 6, 2020
  * ---------------------------------------------------------------------------- */
 
 
@@ -14,7 +14,8 @@
 #define IBEX_AFFINE_AF2_H_
 
 
-#include "ibex_Affine.h"
+#include "ibex_Affine2_fAF2.h"
+#include "ibex_AffineMain.h"
 #include <iostream>
 #include <cassert>
