+++ src/arithmetic/ibex_Affine3_fAFFullI.h
@@ -1,11 +1,11 @@
 /* ============================================================================
- * D Y N I B E X - Definition of the Affine3 class based on fAFFull version 1
+ * I B E X - Definition of the Affine3 class based on fAFFull version 1
  * ============================================================================
- * Copyright   : ENSTA ParisTech
+ * Copyright   : ENSTA Bretagne
  * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
+ *               See the file LICENSE.
  *
- * Author(s)   : Julien Alexandre dit Sandretto and Alexandre Chapoutot
+ * Author(s)   : Jordan Ninin, Julien Alexandre dit Sandretto and Alexandre Chapoutot
  * Created     : Jul 18, 2014
  * Sponsored   : This research benefited from the support of the "Chair Complex Systems Engineering - Ecole Polytechnique, THALES, DGA, FX, DASSAULT AVIATION, DCNS Research, ENSTA ParisTech, Telecom ParisTech, Fondation ParisTech and FDO ENSTA"
  * ---------------------------------------------------------------------------- */
@@ -21,6 +21,18 @@
 
 
 namespace ibex {
+
+
+class AF_fAFFullI;
+
+//typedef AF_fAF1  AF_Other;
+//typedef AF_fAF2  AF_Other;
+//typedef AF_fAF2_fma  AF_Other;
+//typedef AF_iAF  AF_Other;
+//typedef AF_sAF  AF_Other;
+//typedef AF_No  AF_Other;
+typedef AF_fAFFullI AF_Other;
+
 
 template<class T>  class AffineMain;
 template<class T>  class AffineVarMain;
