+++ src/arithmetic/ibex_AffineVar.cpp
@@ -3,13 +3,13 @@
  * ============================================================================
  * Copyright   : ENSTA Bretagne (FRANCE)
  * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
+ *               See the file LICENCE.
  *
  * Author(s)   : Jordan Ninin
- * Created     : Jul 16, 2013
+ * Created     : June 6, 2020
  * ---------------------------------------------------------------------------- */
 #include "ibex_AffineVar.h"
-#include "ibex_Affine.h"
+#include "ibex_AffineMain.h"
 #include <iostream>
 #include <cassert>
