--- src/function/ibex_AffineEval.cpp
@@ -1,21 +1,0 @@
-/* ============================================================================
- * I B E X - ibex_AffineEval.cpp
- * ============================================================================
- * License     : This program can be distributed under the terms of the GNU LGPL.
- *               See the file COPYING.LESSER.
- *
- * Author(s)   : Jordan Ninin
- * Created     : April 08, 2013
- * ---------------------------------------------------------------------------- */
-
-
-#include "ibex_AffineEval.h"
-#include "ibex_Function.h"
-//#include <stdio.h>
-
-namespace ibex {
-
-
-
-} // namespace ibex
-
