+++ src/function/ibex_AffineEval.h
@@ -12,15 +12,14 @@
 #define __IBEX_AFFINE_EVAL_H__
 
 #include "ibex_Function.h"
-#include "ibex_AffineVar.h"
-#include "ibex_AffineMatrix.h"
-#include "ibex_AffineDomain.h"
+#include "ibex_Affine.h"
 #include "ibex_FwdAlgorithm.h"
 #include "ibex_Domain.h"
 #include "ibex_NodeMap.h"
 
 
 namespace ibex {
+template<class T> class AffineEval ;
 
 typedef AffineEval<AF_Default> Affine2Eval;
 typedef AffineEval<AF_Other>  Affine3Eval;
